+++ packages/frameworks/react/src/number-format/stories/number-format.stories.tsx
@@ -1,0 +1,18 @@
+import type { Meta } from '@storybook/react'
+import { NumberFormat } from '../number-format'
+
+const meta: Meta = {
+  title: 'Components / Number Format',
+}
+
+export default meta
+
+export const Basic = () => {
+  return (
+    <div>
+      <span>
+        <NumberFormat value={1204} unit="centimeter" />
+      </span>
+    </div>
+  )
+}