+++ packages/frameworks/solid/src/number-format/index.ts
@@ -1,0 +1,2 @@
+export { NumberFormat } from './number-format'
+export type { NumberFormatProps } from './number-format'