+++ packages/frameworks/vue/src/byte-format/index.ts
@@ -1,0 +1,2 @@
+export { ByteFormat } from './byte-format'
+export type { ByteFormatProps } from './byte-format'