+++ packages/frameworks/solid/CHANGELOG.md
@@ -20,6 +20,7 @@ description: All notable changes to this project will be documented in this file
 ### Added
 
 - Added `Collapsible` component.
+- Added new `ByteFormat` and `NumberFormat` components.
 
 ### Changed